select_highest_overlap

+++ ultralytics/yolo/utils/tal.py
@@ -45,7 +45,7 @@ def select_highest_overlaps(mask_pos, overlaps, n_max_boxes):
         mask_multi_gts = (fg_mask.unsqueeze(1) > 1).repeat([1, n_max_boxes, 1])  # (b, n_max_boxes, h*w)
         max_overlaps_idx = overlaps.argmax(1)  # (b, h*w)
 
-        is_max_overlaps = torch.zeros(mask_pos.shape, dtype=torch.int8, device=mask_pos.device)
+        is_max_overlaps = torch.zeros(mask_pos.shape, dtype=mask_pos.dtype, device=mask_pos.device)
         is_max_overlaps.scatter_(1, max_overlaps_idx.unsqueeze(1), 1)
 
         mask_pos = torch.where(mask_multi_gts, is_max_overlaps, mask_pos).float()  # (b, n_max_boxes, h*w)
